requirements.txt
Django==1.5.4
PIL==1.1.7
South==0.8.2
argparse==1.2.1
django-jquery==1.9.1
django-modeltranslation==0.7.1
psycopg2==2.5.1
six==1.4.1
wsgiref==0.1.2
